RAGKnowledgeBase/requirements.txt
langchain==0.3.25
langchain-openai==0.3.12
langchain-ollama==0.3.3
langchain-chroma==0.2.4
langchain-community==0.3.24
chromadb>=1.0.9
python-dotenv==1.1.0
streamlit==1.45.0
pypdf==5.4.0
flashrank>=0.2.9
